# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/spotify/top/{type}": {
            "get": {
                "description": "GET https://api.spotify.com/v1/me/top/:type.\n* Responds with `items` in JSON representing a list of artists or tracks.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spotify"
                ],
                "summary": "Get top items by type (artists or tracks)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of item to get, either `artists` or `tracks`",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time range to query for top items, either `short_term`, `medium_term`, or `long_term` (default: `medium_term`)",
                        "name": "time_range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}